move out applicationMap from the router
